DX-1071: Move spec helpers to helpers directory
