+++ .docker/entrypoint/.rspec
@@ -1,1 +1,1 @@
---require spec_helper
+--require "helpers/spec_helper"
