+++ .docker/entrypoint/spec/helpers/spec_logger.rb
